Revert "simplify windows service functions and fix api endpoints"
