+++ internal/websockets/client.go
@@ -7,9 +7,11 @@ import (
 	"crypto/tls"
 	"encoding/base64"
 	"fmt"
+	"log"
 	"net/http"
 	"net/url"
 	"os"
+	"os/signal"
 	"sync"
 	"time"
 
@@ -34,15 +36,45 @@ type WSClient struct {
 	Headers         http.Header
 	CommandListener func(*websocket.Conn, Message)
 
-	ctx        context.Context
-	cancel     context.CancelFunc
-	conn       *websocket.Conn
-	send       chan Message
-	mutex      sync.RWMutex
-	writeMutex sync.Mutex
+	ctx         context.Context
+	cancel      context.CancelFunc
+	conn        *websocket.Conn
+	send        chan Message
+	reconnect   chan struct{}
+	mutex       sync.RWMutex
+	writeMutex  sync.Mutex
+	isConnected bool
+	closeOnce   sync.Once
+}
 
-	disconnected chan struct{}
-	connected    bool
+type exponentialBackoff struct {
+	current    time.Duration
+	initial    time.Duration
+	max        time.Duration
+	multiplier float64
+}
+
+func newExponentialBackoff(initial, max time.Duration) *exponentialBackoff {
+	return &exponentialBackoff{
+		initial:    initial,
+		max:        max,
+		current:    initial,
+		multiplier: 2.0,
+	}
+}
+
+func (b *exponentialBackoff) NextBackOff() time.Duration {
+	defer func() {
+		b.current = time.Duration(float64(b.current) * b.multiplier)
+		if b.current > b.max {
+			b.current = b.max
+		}
+	}()
+	return b.current
+}
+
+func (b *exponentialBackoff) Reset() {
+	b.current = b.initial
 }
 
 func NewWSClient(commandListener func(*websocket.Conn, Message)) (*WSClient, error) {
@@ -70,25 +102,88 @@ func NewWSClient(commandListener func(*websocket.Conn, Message)) (*WSClient, err
 		ctx:             ctx,
 		cancel:          cancel,
 		send:            make(chan Message, maxSendBuffer),
-		disconnected:    make(chan struct{}),
+		reconnect:       make(chan struct{}, 1),
 	}
+
+	go client.connectionManager()
+	go client.handleSignals()
 
 	return client, nil
 }
 
-func (c *WSClient) Connect() error {
-	c.mutex.Lock()
-	if c.conn != nil {
-		c.conn.Close()
-		c.conn = nil
-	}
+func (c *WSClient) connectionManager() {
+	backoff := newExponentialBackoff(initialBackoff, reconnectTimeout)
 
-	select {
-	case <-c.disconnected:
-		c.disconnected = make(chan struct{})
-	default:
+	for {
+		select {
+		case <-c.ctx.Done():
+			return
+		default:
+			if err := c.connect(); err != nil {
+				delay := backoff.NextBackOff()
+				log.Printf("Connection failed: %v. Retrying in %v...", err, delay)
+
+				select {
+				case <-time.After(delay):
+					continue // Continue trying to reconnect
+				case <-c.ctx.Done():
+					return
+				}
+			}
+
+			backoff.Reset()
+
+			connCtx, connCancel := context.WithCancel(c.ctx)
+			wg := &sync.WaitGroup{}
+			wg.Add(2)
+
+			errChan := make(chan error, 2)
+
+			go func() {
+				defer wg.Done()
+				c.readPump(connCtx, errChan)
+			}()
+
+			go func() {
+				defer wg.Done()
+				c.writePump(connCtx, errChan)
+			}()
+
+			var err error
+			select {
+			case err = <-errChan:
+				log.Printf("Connection error: %v. Will reconnect...", err)
+			case <-c.ctx.Done():
+				err = c.ctx.Err()
+			}
+
+			connCancel()
+			wg.Wait()
+
+			c.mutex.Lock()
+			if c.conn != nil {
+				c.conn.Close()
+				c.conn = nil
+				c.isConnected = false
+			}
+			c.mutex.Unlock()
+
+			if err == context.Canceled {
+				return
+			}
+
+			continue
+		}
 	}
-	c.mutex.Unlock()
+}
+
+func (c *WSClient) connect() error {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+
+	if c.isConnected {
+		return nil
+	}
 
 	dialer := websocket.Dialer{
 		TLSClientConfig:  &tls.Config{InsecureSkipVerify: true},
@@ -115,41 +210,15 @@ func (c *WSClient) Connect() error {
 		return fmt.Errorf("init message failed: %w", err)
 	}
 
-	c.mutex.Lock()
-	c.connected = true
 	c.conn = conn
-	c.mutex.Unlock()
-
-	// Start read and write pumps
-	errChan := make(chan error, 2)
-	go c.readPump(errChan)
-	go c.writePump(errChan)
-
-	// Return any immediate errors
-	select {
-	case err := <-errChan:
-		return err
-	case <-time.After(time.Second):
-		return nil
-	}
+	c.isConnected = true
+	return nil
 }
 
-func (c *WSClient) handleDisconnect() {
-	c.mutex.Lock()
-	if c.connected {
-		c.connected = false
-		close(c.disconnected)
-	}
-	c.mutex.Unlock()
-}
-
-func (c *WSClient) readPump(errChan chan<- error) {
+func (c *WSClient) readPump(ctx context.Context, errChan chan<- error) {
 	defer func() {
-		c.handleDisconnect()
-		c.conn.Close()
-
 		if r := recover(); r != nil {
-			errChan <- fmt.Errorf("read pump panic: %v\n", r)
+			errChan <- fmt.Errorf("read pump panic: %v", r)
 		}
 	}()
 
@@ -162,7 +231,7 @@ func (c *WSClient) readPump(errChan chan<- error) {
 
 	for {
 		select {
-		case <-c.ctx.Done():
+		case <-ctx.Done():
 			return
 		default:
 			var msg Message
@@ -173,25 +242,18 @@ func (c *WSClient) readPump(errChan chan<- error) {
 			}
 
 			if c.CommandListener != nil {
-				func() {
-					defer func() {
-						if r := recover(); r != nil {
-							errChan <- fmt.Errorf("command listener panic: %v\n", r)
-						}
-					}()
-					c.CommandListener(c.conn, msg)
-				}()
+				c.CommandListener(c.conn, msg)
 			}
 		}
 	}
 }
 
-func (c *WSClient) writePump(errChan chan<- error) {
+func (c *WSClient) writePump(ctx context.Context, errChan chan<- error) {
 	ticker := time.NewTicker(clientPingPeriod)
 	defer func() {
 		ticker.Stop()
 		if r := recover(); r != nil {
-			errChan <- fmt.Errorf("write pump panic: %v\n", r)
+			errChan <- fmt.Errorf("write pump panic: %v", r)
 		}
 	}()
 
@@ -224,49 +286,15 @@ func (c *WSClient) writePump(errChan chan<- error) {
 				return
 			}
 
-		case <-c.ctx.Done():
+		case <-ctx.Done():
 			return
 		}
 	}
 }
 
-func (c *WSClient) Close() error {
-	c.cancel()
-	c.mutex.Lock()
-	defer c.mutex.Unlock()
-
-	if c.conn != nil {
-		// Best effort to send close message
-		c.writeMutex.Lock()
-		c.conn.SetWriteDeadline(time.Now().Add(writeTimeout))
-		c.conn.WriteMessage(websocket.CloseMessage, websocket.FormatCloseMessage(websocket.CloseNormalClosure, ""))
-		c.writeMutex.Unlock()
-
-		// Give time for close message to be sent
-		time.Sleep(time.Second)
-		c.conn.Close()
-		c.conn = nil
-	}
-
-	close(c.send)
-	return nil
-}
-
-func (c *WSClient) Wait() <-chan struct{} {
-	c.mutex.RLock()
-	defer c.mutex.RUnlock()
-	return c.disconnected
-}
-
-func (c *WSClient) IsConnected() bool {
-	c.mutex.RLock()
-	defer c.mutex.RUnlock()
-	return c.connected
-}
-
 func (c *WSClient) SendMessage(msg Message) error {
 	c.mutex.RLock()
-	if c.conn == nil {
+	if !c.isConnected {
 		c.mutex.RUnlock()
 		return fmt.Errorf("not connected")
 	}
@@ -280,6 +308,43 @@ func (c *WSClient) SendMessage(msg Message) error {
 	case <-c.ctx.Done():
 		return fmt.Errorf("client closed")
 	}
+}
+
+func (c *WSClient) handleSignals() {
+	interrupt := make(chan os.Signal, 1)
+	signal.Notify(interrupt, os.Interrupt)
+
+	select {
+	case <-interrupt:
+		c.Close()
+	case <-c.ctx.Done():
+		signal.Stop(interrupt)
+		close(interrupt)
+	}
+}
+
+func (c *WSClient) Close() {
+	c.closeOnce.Do(func() {
+		c.cancel() // Cancel context first to stop all goroutines
+
+		c.mutex.Lock()
+		defer c.mutex.Unlock()
+
+		if c.conn != nil {
+			// Best effort to send close message
+			c.writeMutex.Lock()
+			c.conn.SetWriteDeadline(time.Now().Add(writeTimeout))
+			c.conn.WriteMessage(websocket.CloseMessage, websocket.FormatCloseMessage(websocket.CloseNormalClosure, ""))
+			c.writeMutex.Unlock()
+
+			time.Sleep(time.Second) // Give time for close message to be sent
+			c.conn.Close()
+			c.conn = nil
+			c.isConnected = false
+		}
+
+		close(c.send)
+	})
 }
 
 func validateRegistryValue(value string, maxLength int) (string, error) {
